@@ -12,6 +12,8 @@
 # language governing permissions and limitations under the License.
 import os
 
+from ibm_botocore.docs import DEPRECATED_SERVICE_NAMES
+
 from ibm_boto3.docs.service import ServiceDocumenter
 
 
@@ -31,8 +33,17 @@ def generate_docs(root_dir, session):
     if not os.path.exists(services_doc_path):
         os.makedirs(services_doc_path)
 
-    for service_name in session.get_available_services():
-        docs = ServiceDocumenter(service_name, session).document_service()
+    # Prevents deprecated service names from being generated in docs.
+    available_services = [
+        service
+        for service in session.get_available_services()
+        if service not in DEPRECATED_SERVICE_NAMES
+    ]
+
+    for service_name in available_services:
+        docs = ServiceDocumenter(
+            service_name, session, services_doc_path
+        ).document_service()
         service_doc_path = os.path.join(
             services_doc_path, service_name + '.rst'
         )
