@@ -26,12 +26,13 @@ class ServiceDocumenter(BaseServiceDocumenter):
     # The path used to find examples
     EXAMPLE_PATH = os.path.join(os.path.dirname(ibm_boto3.__file__), 'examples')
 
-    def __init__(self, service_name, session):
+    def __init__(self, service_name, session, root_docs_path):
         super().__init__(
             service_name=service_name,
             # I know that this is an internal attribute, but the ibm_botocore session
             # is needed to load the paginator and waiter models.
             session=session._session,
+            root_docs_path=root_docs_path,
         )
         self._ibm_boto3_session = session
         self._client = self._ibm_boto3_session.client(service_name)
@@ -40,14 +41,18 @@ class ServiceDocumenter(BaseServiceDocumenter):
             self._service_resource = self._ibm_boto3_session.resource(service_name)
         self.sections = [
             'title',
-            'table-of-contents',
             'client',
             'paginators',
             'waiters',
-            'service-resource',
             'resources',
             'examples',
+            'context-params',
         ]
+        self._root_docs_path = root_docs_path
+        self._USER_GUIDE_LINK = (
+            'https://ibm_boto3.amazonaws.com/'
+            'v1/documentation/api/latest/guide/resources.html'
+        )
 
     def document_service(self):
         """Documents an entire service.
@@ -58,17 +63,15 @@ class ServiceDocumenter(BaseServiceDocumenter):
             self._service_name, section_names=self.sections, target='html'
         )
         self.title(doc_structure.get_section('title'))
-        self.table_of_contents(doc_structure.get_section('table-of-contents'))
 
         self.client_api(doc_structure.get_section('client'))
         self.paginator_api(doc_structure.get_section('paginators'))
         self.waiter_api(doc_structure.get_section('waiters'))
         if self._service_resource:
-            self._document_service_resource(
-                doc_structure.get_section('service-resource')
-            )
-            self._document_resources(doc_structure.get_section('resources'))
+            self.resource_section(doc_structure.get_section('resources'))
         self._document_examples(doc_structure.get_section('examples'))
+        context_params_section = doc_structure.get_section('context-params')
+        self.client_context_params(context_params_section)
         return doc_structure.flush_structure()
 
     def client_api(self, section):
@@ -78,12 +81,57 @@ class ServiceDocumenter(BaseServiceDocumenter):
         except DataNotFoundError:
             pass
 
-        Boto3ClientDocumenter(self._client, examples).document_client(section)
+        Boto3ClientDocumenter(
+            self._client, self._root_docs_path, examples
+        ).document_client(section)
+
+    def resource_section(self, section):
+        section.style.h2('Resources')
+        section.style.new_line()
+        section.write(
+            'Resources are available in ibm_boto3 via the '
+            '``resource`` method. For more detailed instructions '
+            'and examples on the usage of resources, see the '
+            'resources '
+        )
+        section.style.external_link(
+            title='user guide',
+            link=self._USER_GUIDE_LINK,
+        )
+        section.write('.')
+        section.style.new_line()
+        section.style.new_line()
+        section.write('The available resources are:')
+        section.style.new_line()
+        section.style.toctree()
+        self._document_service_resource(section)
+        self._document_resources(section)
 
     def _document_service_resource(self, section):
+        # Create a new DocumentStructure for each Service Resource and add contents.
+        service_resource_doc = DocumentStructure(
+            'service-resource', target='html'
+        )
+        breadcrumb_section = service_resource_doc.add_new_section('breadcrumb')
+        breadcrumb_section.style.ref(
+            self._client.__class__.__name__, f'../../{self._service_name}'
+        )
+        breadcrumb_section.write(' / Resource / ServiceResource')
         ServiceResourceDocumenter(
-            self._service_resource, self._session
-        ).document_resource(section)
+            self._service_resource, self._session, self._root_docs_path
+        ).document_resource(service_resource_doc)
+        # Write collections in individual/nested files.
+        # Path: <root>/reference/services/<service>/<resource_name>/<collection_name>.rst
+        resource_name = self._service_resource.meta.resource_model.name
+        if resource_name == self._service_name:
+            resource_name = 'service-resource'
+        service_resource_dir_path = os.path.join(
+            self._root_docs_path,
+            f'{self._service_name}',
+            f'{resource_name.lower()}',
+        )
+        service_resource_doc.write_to_file(service_resource_dir_path, 'index')
+        section.style.tocitem(f'{self._service_name}/{resource_name}/index')
 
     def _document_resources(self, section):
         temp_identifier_value = 'foo'
@@ -113,8 +161,31 @@ class ServiceDocumenter(BaseServiceDocumenter):
             for _ in identifiers:
                 args.append(temp_identifier_value)
             resource = resource_cls(*args, client=self._client)
-            ResourceDocumenter(resource, self._session).document_resource(
-                section.add_new_section(resource.meta.resource_model.name)
+            # Create a new DocumentStructure for each Resource and add contents.
+            resource_name = resource.meta.resource_model.name.lower()
+            resource_doc = DocumentStructure(resource_name, target='html')
+            breadcrumb_section = resource_doc.add_new_section('breadcrumb')
+            breadcrumb_section.style.ref(
+                self._client.__class__.__name__, f'../../{self._service_name}'
+            )
+            breadcrumb_section.write(
+                f' / Resource / {resource.meta.resource_model.name}'
+            )
+            ResourceDocumenter(
+                resource, self._session, self._root_docs_path
+            ).document_resource(
+                resource_doc.add_new_section(resource.meta.resource_model.name)
+            )
+            # Write collections in individual/nested files.
+            # Path: <root>/reference/services/<service>/<resource_name>/<index>.rst
+            service_resource_dir_path = os.path.join(
+                self._root_docs_path,
+                f'{self._service_name}',
+                f'{resource_name}',
+            )
+            resource_doc.write_to_file(service_resource_dir_path, 'index')
+            section.style.tocitem(
+                f'{self._service_name}/{resource_name}/index'
             )
 
     def _get_example_file(self):
@@ -126,9 +197,6 @@ class ServiceDocumenter(BaseServiceDocumenter):
         examples_file = self._get_example_file()
         if os.path.isfile(examples_file):
             section.style.h2('Examples')
-            section.style.new_line()
-            section.write(".. contents::\n    :local:\n    :depth: 1")
-            section.style.new_line()
             section.style.new_line()
             with open(examples_file) as f:
                 section.write(f.read())
