@@ -10,10 +10,13 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
+import os
+
 from ibm_botocore import xform_name
+from ibm_botocore.docs.bcdoc.restdoc import DocumentStructure
 from ibm_botocore.utils import get_service_module_name
 
-from ibm_boto3.docs.base import BaseDocumenter
+from ibm_boto3.docs.base import NestedDocumenter
 from ibm_boto3.docs.utils import (
     add_resource_type_overview,
     get_identifier_args_for_signature,
@@ -22,7 +25,7 @@ from ibm_boto3.docs.utils import (
 )
 
 
-class SubResourceDocumenter(BaseDocumenter):
+class SubResourceDocumenter(NestedDocumenter):
     def document_sub_resources(self, section):
         add_resource_type_overview(
             section=section,
@@ -43,11 +46,34 @@ class SubResourceDocumenter(BaseDocumenter):
         for sub_resource in sub_resources:
             sub_resource_section = section.add_new_section(sub_resource.name)
             sub_resources_list.append(sub_resource.name)
+            # Create a new DocumentStructure for each sub_resource and add contents.
+            sub_resource_doc = DocumentStructure(
+                sub_resource.name, target='html'
+            )
+            breadcrumb_section = sub_resource_doc.add_new_section('breadcrumb')
+            breadcrumb_section.style.ref(self._resource_class_name, 'index')
+            breadcrumb_section.write(f' / Sub-Resource / {sub_resource.name}')
+            sub_resource_doc.add_title_section(sub_resource.name)
+            sub_resource_section = sub_resource_doc.add_new_section(
+                sub_resource.name,
+                context={'qualifier': f'{self.class_name}.'},
+            )
             document_sub_resource(
                 section=sub_resource_section,
                 resource_name=self._resource_name,
                 sub_resource_model=sub_resource,
                 service_model=self._service_model,
+            )
+
+            # Write sub_resources in individual/nested files.
+            # Path: <root>/reference/services/<service>/<resource_name>/<sub_resource_name>.rst
+            sub_resources_dir_path = os.path.join(
+                self._root_docs_path,
+                f'{self._service_name}',
+                f'{self._resource_sub_path}',
+            )
+            sub_resource_doc.write_to_file(
+                sub_resources_dir_path, sub_resource.name
             )
 
 
@@ -78,8 +104,11 @@ def document_sub_resource(
 
     if include_signature:
         signature_args = get_identifier_args_for_signature(identifiers_needed)
+        full_sub_resource_name = (
+            f"{section.context.get('qualifier', '')}{sub_resource_model.name}"
+        )
         section.style.start_sphinx_py_method(
-            sub_resource_model.name, signature_args
+            full_sub_resource_name, signature_args
         )
 
     method_intro_section = section.add_new_section('method-intro')
