@@ -17,8 +17,11 @@ from tests.functional.docs import BaseDocsFunctionalTests
 
 class TestS3Customizations(BaseDocsFunctionalTests):
     def setUp(self):
+        super().setUp()
         self.documenter = ServiceDocumenter(
-            's3', session=Session(region_name='us-east-1')
+            's3',
+            session=Session(region_name='us-east-1'),
+            root_docs_path=self.root_services_path,
         )
         self.generated_contents = self.documenter.document_service()
         self.generated_contents = self.generated_contents.decode('utf-8')
@@ -26,11 +29,33 @@ class TestS3Customizations(BaseDocsFunctionalTests):
     def test_file_transfer_methods_are_documented(self):
         self.assert_contains_lines_in_order(
             [
-                '.. py:class:: S3.Client',
-                '  *   :py:meth:`~S3.Client.download_file`',
-                '  *   :py:meth:`~S3.Client.upload_file`',
-                '  .. py:method:: download_file(',
-                '  .. py:method:: upload_file(',
+                '=========',
+                'Resources',
+                '=========',
+                'The available resources are:',
+                '  s3/service-resource/index',
             ],
             self.generated_contents,
+        )
+        self.assert_contains_lines_in_order(
+            [
+                '.. py:class:: S3.Client',
+                '  s3/client/download_file',
+                '  s3/client/upload_file',
+            ],
+            self.generated_contents,
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'download_file',
+                '.. py:method:: S3.Client.download_file(',
+            ],
+            self.get_nested_file_contents('s3', 'client', 'download_file'),
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'upload_file',
+                '.. py:method:: S3.Client.upload_file(',
+            ],
+            self.get_nested_file_contents('s3', 'client', 'upload_file'),
         )
