@@ -60,7 +60,7 @@ class TestResourceDocstrings(BaseDocsTest):
                     '  Calls :py:meth:`MyService.Client.sample_operation` to update '
                     'the attributes of the Sample resource'
                 ),
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    sample.load()',
                 '  :returns: None',
@@ -136,7 +136,7 @@ class TestResourceDocstrings(BaseDocsTest):
                     '    Creates an iterable of all Sample resources in the '
                     'collection.'
                 ),
-                '    **Request Syntax** ',
+                '    **Request Syntax**',
                 '    ::',
                 '      sample_iterator = myservice.samples.all()',
                 '    :rtype: list(:py:class:`myservice.Sample`)',
@@ -151,7 +151,7 @@ class TestResourceDocstrings(BaseDocsTest):
         collection_method_docstring = mock_stdout.getvalue()
         self.assert_contains_lines_in_order(
             [
-                '    **Request Syntax** ',
+                '    **Request Syntax**',
                 '    ::',
                 '      sample_iterator = myservice.samples.filter(',
                 "          Foo='string',",
@@ -173,7 +173,7 @@ class TestResourceDocstrings(BaseDocsTest):
         collection_method_docstring = mock_stdout.getvalue()
         self.assert_contains_lines_in_order(
             [
-                '    **Request Syntax** ',
+                '    **Request Syntax**',
                 '    ::',
                 '      sample_iterator = myservice.samples.limit(',
                 '          count=123',
@@ -195,7 +195,7 @@ class TestResourceDocstrings(BaseDocsTest):
         collection_method_docstring = mock_stdout.getvalue()
         self.assert_contains_lines_in_order(
             [
-                '    **Request Syntax** ',
+                '    **Request Syntax**',
                 '    ::',
                 '      sample_iterator = myservice.samples.page_size(',
                 '          count=123',
@@ -222,7 +222,7 @@ class TestResourceDocstrings(BaseDocsTest):
                     '    Creates an iterable of all Sample resources in the '
                     'collection.'
                 ),
-                '    **Request Syntax** ',
+                '    **Request Syntax**',
                 '    ::',
                 '      sample_iterator = myservice.samples.all()',
                 '    :rtype: list(:py:class:`myservice.Sample`)',
@@ -237,7 +237,7 @@ class TestResourceDocstrings(BaseDocsTest):
         batch_action_docstring = mock_stdout.getvalue()
         self.assert_contains_lines_in_order(
             [
-                '    **Request Syntax** ',
+                '    **Request Syntax**',
                 '    ::',
                 '      response = myservice.samples.operate(',
                 "          Foo='string',",
@@ -248,15 +248,15 @@ class TestResourceDocstrings(BaseDocsTest):
                 '    :type Bar: string',
                 '    :param Bar: Documents Bar',
                 '    :rtype: dict',
-                '    :returns: ',
-                '      **Response Syntax** ',
+                '    :returns:',
+                '      **Response Syntax**',
                 '      ::',
                 '        {',
                 "            'Foo': 'string',",
                 "            'Bar': 'string'",
                 '        }',
-                '      **Response Structure** ',
-                '      - *(dict) --* ',
+                '      **Response Structure**',
+                '      - *(dict) --*',
                 '        - **Foo** *(string) --* Documents Foo',
                 '        - **Bar** *(string) --* Documents Bar',
             ],
@@ -272,11 +272,11 @@ class TestResourceDocstrings(BaseDocsTest):
                 (
                     '    Waits until this Sample is complete. This method calls '
                     ':py:meth:`MyService.Waiter.sample_operation_complete.wait` '
-                    'which polls. :py:meth:`MyService.Client.sample_operation` every '
+                    'which polls :py:meth:`MyService.Client.sample_operation` every '
                     '15 seconds until a successful state is reached. An error '
                     'is returned after 40 failed checks.'
                 ),
-                '    **Request Syntax** ',
+                '    **Request Syntax**',
                 '    ::',
                 '      sample.wait_until_complete(',
                 "          Bar='string'",
