@@ -42,7 +42,7 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
             [
                 '.. py:method:: foo(**kwargs)',
                 '  This describes the foo method.',
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    response = myservice.foo(',
                 "        Foo='string',",
@@ -54,13 +54,13 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
                 '  :param Bar: Documents Bar',
                 '  :rtype: dict',
                 '  :returns: ',
-                '    **Response Syntax** ',
+                '    **Response Syntax**',
                 '    ::',
                 '      {',
                 "          'Foo': 'string',",
                 "          'Bar': 'string'",
                 '      }',
-                '    **Response Structure** ',
+                '    **Response Structure**',
                 '    - *(dict) --* ',
                 '      - **Foo** *(string) --* Documents Foo',
                 '      - **Bar** *(string) --* Documents Bar',
@@ -97,7 +97,7 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
             [
                 '.. py:method:: foo(**kwargs)',
                 '  This describes the foo method.',
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    sample = myservice.foo(',
                 "        Foo='string',",
@@ -127,7 +127,7 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
             [
                 '.. py:method:: foo(**kwargs)',
                 '  This describes the foo method.',
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    samples = myservice.foo(',
                 "        Foo='string',",
@@ -162,7 +162,7 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
             [
                 '.. py:method:: foo(**kwargs)',
                 '  This describes the foo method.',
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    response = myservice.foo(',
                 "        Foo='string',",
@@ -177,13 +177,13 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
                 '  :param Biz: biz docs',
                 '  :rtype: dict',
                 '  :returns: ',
-                '    **Response Syntax** ',
+                '    **Response Syntax**',
                 '    ::',
                 '      {',
                 "          'Foo': 'string',",
                 "          'Bar': 'string'",
                 '      }',
-                '    **Response Structure** ',
+                '    **Response Structure**',
                 '    - *(dict) --* ',
                 '      - **Foo** *(string) --* Documents Foo',
                 '      - **Bar** *(string) --* Documents Bar',
@@ -210,7 +210,7 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
             [
                 '.. py:method:: foo(**kwargs)',
                 '  This describes the foo method.',
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    response = myservice.foo(',
                 "        Foo='string',",
@@ -222,14 +222,14 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
                 '  :param Bar: Documents Bar',
                 '  :rtype: dict',
                 '  :returns: ',
-                '    **Response Syntax** ',
+                '    **Response Syntax**',
                 '    ::',
                 '      {',
                 "          'Foo': 'string',",
                 "          'Bar': 'string',",
                 "          'Biz': 'string'",
                 '      }',
-                '    **Response Structure** ',
+                '    **Response Structure**',
                 '    - *(dict) --* ',
                 '      - **Foo** *(string) --* Documents Foo',
                 '      - **Bar** *(string) --* Documents Bar',
@@ -252,7 +252,7 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
             [
                 '.. py:method:: foo(**kwargs)',
                 '  This describes the foo method.',
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    response = myservice.foo(',
                 "        Foo='string',",
@@ -261,13 +261,13 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
                 '  :param Foo: Documents Foo',
                 '  :rtype: dict',
                 '  :returns: ',
-                '    **Response Syntax** ',
+                '    **Response Syntax**',
                 '    ::',
                 '      {',
                 "          'Foo': 'string',",
                 "          'Bar': 'string'",
                 '      }',
-                '    **Response Structure** ',
+                '    **Response Structure**',
                 '    - *(dict) --* ',
                 '      - **Foo** *(string) --* Documents Foo',
                 '      - **Bar** *(string) --* Documents Bar',
@@ -292,7 +292,7 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
             [
                 '.. py:method:: foo(**kwargs)',
                 '  This describes the foo method.',
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    response = myservice.foo(',
                 "        Foo='string',",
@@ -304,12 +304,12 @@ class TestDocumentModelDrivenResourceMethod(BaseDocsTest):
                 '  :param Bar: Documents Bar',
                 '  :rtype: dict',
                 '  :returns: ',
-                '    **Response Syntax** ',
+                '    **Response Syntax**',
                 '    ::',
                 '      {',
                 "          'Foo': 'string'",
                 '      }',
-                '    **Response Structure** ',
+                '    **Response Structure**',
                 '    - *(dict) --* ',
                 '      - **Foo** *(string) --* Documents Foo',
             ]
