@@ -18,7 +18,7 @@ class TestResourceDocumenter(BaseDocsTest):
     def test_document_resource(self):
         resource = self.resource.Sample('mysample')
         resource_documenter = ResourceDocumenter(
-            resource, self.ibm_botocore_session
+            resource, self.ibm_botocore_session, self.root_services_path
         )
         resource_documenter.document_resource(self.doc_structure)
         self.assert_contains_lines_in_order(
@@ -26,47 +26,80 @@ class TestResourceDocumenter(BaseDocsTest):
                 '======',
                 'Sample',
                 '======',
+                'Before using anything on this page, please refer to the resources ',
+                ':doc:`user guide <../../../../guide/resources>` for the most recent ',
+                'guidance on using resources.',
                 '.. py:class:: MyService.Sample(name)',
                 '  A resource representing an AWS MyService Sample::',
                 '    import ibm_boto3',
                 "    myservice = ibm_boto3.resource('myservice')",
                 "    sample = myservice.Sample('name')",
-                "  These are the resource's available identifiers:",
-                '  *   :py:attr:`name`',
-                "  These are the resource's available attributes:",
-                '  *   :py:attr:`bar`',
-                '  *   :py:attr:`foo`',
-                "  These are the resource's available actions:",
-                '  *   :py:meth:`load()`',
-                '  *   :py:meth:`operate()`',
-                '  *   :py:meth:`reload()`',
-                "  These are the resource's available waiters:",
-                '  *   :py:meth:`wait_until_complete()`',
-                '  .. rst-class:: admonition-title',
-                '  Identifiers',
-                '  .. py:attribute:: name',
-                '  .. rst-class:: admonition-title',
-                '  Attributes',
-                '  .. py:attribute:: bar',
-                '    - *(string) --* Documents Bar',
-                '  .. py:attribute:: foo',
-                '    - *(string) --* Documents Foo',
-                '  .. rst-class:: admonition-title',
-                '  Actions',
-                '  .. py:method:: load()',
-                '  .. py:method:: operate(**kwargs)',
-                '  .. py:method:: reload()',
-                '  .. rst-class:: admonition-title',
-                '  Waiters',
-                '  .. py:method:: wait_until_complete(**kwargs)',
+                'Identifiers',
+                "These are the resource's available identifiers:",
+                '.. toctree::',
+                '  :maxdepth: 1',
+                '  :titlesonly:',
+                '  name',
+                'Attributes',
+                "These are the resource's available attributes:",
+                '.. toctree::',
+                '  :maxdepth: 1',
+                '  :titlesonly:',
+                '  bar',
+                '  foo',
+                'Actions',
+                "These are the resource's available actions:",
+                '.. toctree::',
+                '  :maxdepth: 1',
+                '  :titlesonly:',
+                '  load',
+                '  operate',
+                '  reload',
+                'Waiters',
+                "These are the resource's available waiters:",
+                '.. toctree::',
+                '  :maxdepth: 1',
+                '  :titlesonly:',
+                '  wait_until_complete',
             ]
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'name',
+                '.. py:attribute:: MyService.Sample.name',
+            ],
+            self.get_nested_service_contents('myservice', 'sample', 'name'),
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'bar',
+                '.. py:attribute:: MyService.Sample.bar',
+                '  - *(string) --* Documents Bar',
+            ],
+            self.get_nested_service_contents('myservice', 'sample', 'bar'),
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'load',
+                '.. py:method:: MyService.Sample.load()',
+            ],
+            self.get_nested_service_contents('myservice', 'sample', 'load'),
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'wait_until_complete',
+                '.. py:method:: MyService.Sample.wait_until_complete(**kwargs)',
+            ],
+            self.get_nested_service_contents(
+                'myservice', 'sample', 'wait_until_complete'
+            ),
         )
 
 
 class TestServiceResourceDocumenter(BaseDocsTest):
     def test_document_resource(self):
         resource_documenter = ServiceResourceDocumenter(
-            self.resource, self.ibm_botocore_session
+            self.resource, self.ibm_botocore_session, self.root_services_path
         )
         resource_documenter.document_resource(self.doc_structure)
         self.assert_contains_lines_in_order(
@@ -74,28 +107,61 @@ class TestServiceResourceDocumenter(BaseDocsTest):
                 '================',
                 'Service Resource',
                 '================',
+                'Before using anything on this page, please refer to the resources ',
+                ':doc:`user guide <../../../../guide/resources>` for the most recent ',
+                'guidance on using resources.',
                 '.. py:class:: MyService.ServiceResource()',
                 '  A resource representing AWS MyService::',
                 '    import ibm_boto3',
                 "    myservice = ibm_boto3.resource('myservice')",
-                "  These are the resource's available actions:",
-                '  *   :py:meth:`sample_operation()`',
-                "  These are the resource's available sub-resources:",
-                '  *   :py:meth:`Sample()`',
-                "  These are the resource's available collections:",
-                '  *   :py:attr:`samples`',
-                '  .. rst-class:: admonition-title',
-                '  Actions',
-                '  .. py:method:: sample_operation(**kwargs)',
-                '  .. rst-class:: admonition-title',
-                '  Sub-resources',
-                '  .. py:method:: Sample(name)',
-                '  .. rst-class:: admonition-title',
-                '  Collections',
-                '  .. py:attribute:: samples',
-                '    .. py:method:: all()',
-                '    .. py:method:: filter(**kwargs)',
-                '    .. py:method:: limit(**kwargs)',
-                '    .. py:method:: page_size(**kwargs)',
+                'Actions',
+                "These are the resource's available actions:",
+                '.. toctree::',
+                '  :maxdepth: 1',
+                '  :titlesonly:',
+                '  sample_operation',
+                'Sub-resources',
+                "These are the resource's available sub-resources:",
+                '.. toctree::',
+                '  :maxdepth: 1',
+                '  :titlesonly:',
+                '  Sample',
+                'Collections',
+                "These are the resource's available collections:",
+                '.. toctree::',
+                '  :maxdepth: 1',
+                '  :titlesonly:',
+                '  samples',
             ]
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'sample_operation',
+                '.. py:method:: MyService.ServiceResource.sample_operation(**kwargs)',
+            ],
+            self.get_nested_service_contents(
+                'myservice', 'service-resource', 'sample_operation'
+            ),
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'Sample',
+                '.. py:method:: MyService.ServiceResource.Sample(name)',
+            ],
+            self.get_nested_service_contents(
+                'myservice', 'service-resource', 'Sample'
+            ),
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'samples',
+                '.. py:attribute:: MyService.ServiceResource.samples',
+                '  .. py:method:: all()',
+                '  .. py:method:: filter(**kwargs)',
+                '  .. py:method:: limit(**kwargs)',
+                '  .. py:method:: page_size(**kwargs)',
+            ],
+            self.get_nested_service_contents(
+                'myservice', 'service-resource', 'samples'
+            ),
         )
