@@ -21,20 +21,29 @@ class TestWaiterResourceDocumenter(BaseDocsTest):
         )
         subresource = self.resource.Sample('mysample')
         waiter_documenter = WaiterResourceDocumenter(
-            subresource, service_waiter_model
+            subresource, service_waiter_model, self.root_services_path
         )
         waiter_documenter.document_resource_waiters(self.doc_structure)
         self.assert_contains_lines_in_order(
             [
-                '.. py:method:: wait_until_complete(**kwargs)',
+                '-------\nWaiters\n-------',
+                'Waiters provide an interface to wait for a resource',
+                ' to reach a specific state.',
+                'For more information about waiters refer to the',
+            ]
+        )
+        self.assert_contains_lines_in_order(
+            [
+                'wait_until_complete',
+                '.. py:method:: MyService.Sample.wait_until_complete(**kwargs)',
                 (
                     '  Waits until this Sample is complete. This method calls '
                     ':py:meth:`MyService.Waiter.sample_operation_complete.wait` '
-                    'which polls. :py:meth:`MyService.Client.sample_operation` '
+                    'which polls :py:meth:`MyService.Client.sample_operation` '
                     'every 15 seconds until a successful state is reached. An '
                     'error is returned after 40 failed checks.'
                 ),
-                '  **Request Syntax** ',
+                '  **Request Syntax**',
                 '  ::',
                 '    sample.wait_until_complete(',
                 "        Bar='string'",
@@ -42,5 +51,8 @@ class TestWaiterResourceDocumenter(BaseDocsTest):
                 '  :type Bar: string',
                 '  :param Bar: Documents Bar',
                 '  :returns: None',
-            ]
+            ],
+            self.get_nested_service_contents(
+                'myservice', 'sample', 'wait_until_complete'
+            ),
         )
